GUENTHER’S MARKET
Task: Make a shopping list application which uses collections to store your items.

What will the application do:

Display a list of at least 8 item names and prices.
Ask the user to enter an item name
If that item exists, display that item and price and add that item and its price to the user’s order.
If that item doesn’t exist, display an error and re-prompt the user.  (Display the menu again if you’d like.)
Ask if they want to add another. Repeat if they do. (Th user can enter an item more than once; we’re not taking quantity at this point.)
When they’re done adding items, display a list of all items ordered with prices in columns.
Display the average price of items ordered.

Build Specifications:

Use a Map to keep track of the menu of items. It should have a String key (for item name) and Double value (for item price).
Use parallel ArrayLists (one of strings, one of doubles) to store the items ordered and their prices.
Write 3 methods to find 1) the average cost of the items ordered and the indexes of the 2) highest and 3) lowest cost items.